+++ dffml/service/dev.py
@@ -63,6 +63,7 @@ if sys.version_info.major == 3 and sys.version_info.minor < 8:
         ("model", "tensorflow"),
         ("model", "tensorflow_hub"),
         ("model", "transformers"),
+        ("model", "vowpalWabbit"),
     ]